+++ src/ballistica/shared/foundation/fatal_error.cc
@@ -37,7 +37,7 @@ void FatalError::ReportFatalError(const std::string& message,
   // immediately in order to get the debugger's attention.
   if (g_core && g_core->core_config().debugger_attached) {
     if (!message.empty()) {
-      printf("FATAL ERROR: %s\n", message.c_str());
+      printf("FATAL ERROR (debugger mode): %s\n", message.c_str());
       fflush(stdout);
     }
     abort();
@@ -72,12 +72,14 @@ void FatalError::ReportFatalError(const std::string& message,
     if (g_core && g_core->platform) {
       core::PlatformStackTrace* trace{g_core->platform->GetStackTrace()};
       if (trace) {
-        std::string tracestr = trace->GetDescription();
+        std::string tracestr = trace->FormatForDisplay();
         if (!tracestr.empty()) {
           logmsg += ("\nCPP-STACK-TRACE-BEGIN:\n" + tracestr
                      + "\nCPP-STACK-TRACE-END");
         }
         delete trace;
+      } else {
+        logmsg += "\n(CPP-STACK-TRACE-UNAVAILABLE)";
       }
     }
   }