+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-### 1.7.20 (build 21071, api 8, 2023-06-07)
+### 1.7.20 (build 21079, api 8, 2023-06-08)
 
 - This seems like a good time for a `refactoring` release in anticipation of
   changes coming in 1.8. Basically this means that a lot of things will be
@@ -299,6 +299,20 @@
   (thanks for the heads- up SEBASTIAN2059)
 - (build 21070) Fixed an issue where teams series would incorrectly end after 1
   round (thanks for the heads up SEBASTIAN2059)
+- (build 21072) Fixed a crash drawing a terrain node with no texture set.
+- (build 21073) Stack traces are now implemented under windows so should show up
+  for fatal errors and whatnot. Also fatal error logging now mentions when stack
+  traces are not available.
+- (build 21074) Added `babase.native_stack_trace()` to fetch native stack traces
+  as strings.
+- (build 21076) Hopefully fixed a 'file in use' error for `_appstate_dump_tb` on
+  windows. Please holler if you are still seeing this. This file gets written
+  for debugging whenever the logic thread remains unresponsive for several
+  seconds.
+- (build 21078) Custom system scripts dirs works again (complete copies of app
+  system scripts living in your mods directory under `sys/$(YOUR_APP_VERSION)`.
+  Tools for creating/destroying these setups are now at `babase.modutils` (they
+  had been placed under bauiv1 but that was just silly).
 
 ### 1.7.19 (build 20997, api 7, 2023-01-19)